Fix stupid bug, loosen PT cuts

+++ python/Configuration.py
@@ -106,13 +106,13 @@ def addTTWAnalyzer(framework, name="ttW", prefix="ttW_"):
             jetsProducer = cms.string('jets'),
             metProducer = cms.string('met'),
 
-            electronPtCut = cms.untracked.double(20),
+            electronPtCut = cms.untracked.double(10),
             electronEtaCut = cms.untracked.double(2.5),
 
-            muonPtCut = cms.untracked.double(20),
+            muonPtCut = cms.untracked.double(10),
             muonEtaCut = cms.untracked.double(2.4),
 
-            jetPtCut = cms.untracked.double(30),
+            jetPtCut = cms.untracked.double(25),
             jetEtaCut = cms.untracked.double(2.5),
             #jetPUID = cms.untracked.double(-9999999),
             jetDRleptonCut = cms.untracked.double(0.3),
